requirements.txt
langchain_core==0.3.51
langchain_ollama==0.3.1
langgraph==0.3.26
streamlit==1.37.1
tavily_python==0.5.1
torch==2.3.0
torch==2.3.0+cu121
